Use size_t for _local_id&co. This removes a zext in address computation which makes the offsets analyzable and coalesceable (unh!).
Cannot commit this to trunk because the poclizer passes assume the size_t
is (now) always 64bits which is not the case for 32bit targets such as
TCE. Need to figure it out from the target we are compiling for.

+++ lib/kernel/get_global_size.c
@@ -21,13 +21,13 @@
    THE SOFTWARE.
 */
 
-extern unsigned int _local_size_x;
-extern unsigned int _local_size_y;
-extern unsigned int _local_size_z;
+extern size_t _local_size_x;
+extern size_t _local_size_y;
+extern size_t _local_size_z;
 
-extern unsigned int _num_groups_x;
-extern unsigned int _num_groups_y;
-extern unsigned int _num_groups_z;
+extern size_t _num_groups_x;
+extern size_t _num_groups_y;
+extern size_t _num_groups_z;
 
 size_t
 get_global_size(unsigned int dimindx)
